Refactor data structure and update coupon/campaign logic
Moved campaigns and categories data from 'store' and 'utils' directories to a new 'data' directory for better organization. Updated imports in CouponCard, CouponForm, Campaigns, and Coupons components/pages to reflect new data paths. This refactor improves code maintainability and separation of concerns.
